*Sighs*
Even with the HTML 'notranslation' fix, for *some* languages this still fails. The problem appears to be on the Google end. For some languages, given the string ...

"Some text <span class="notranslate">NO_TRANSLATE</span> and more text  <span class="notranslate">NO_TRANSLATE</span> the end"

.. you end up with ..

"Some text <span class="notranslate">NO_TRANSLATE</span> and more text NO_TRANSLATE the end"

To combat this, multiple 'no translate strings' can be specified. In this case, only the FIRST string is used in the actual text sent to Google, but all strings are checked for in the response.